Tests/LibWeb/Text/expected/scroll-does-not-rebuild-accumulated-visual-context-tree.txt

scrollTop: 100
scrollLeft: 50
bounding rect y: 8
builds after pure scroll: 0
